Start openmpi@3.1.6 % gcc@10.2.0 dep chain for expanse/0.15.4/cpu
At present, this dependency chain is only intended to be used for sprng,
a distributied random number generator. It is an older package and
requires an MPI distribution that supports cxx bindings, which were
deprecated in the OpenMPI 4.X.X series. Note, however, we expect this
older MPI distribution may be used to support other older packages as
needed.