Start openmpi@3.1.4 % gcc@10.2.0 dep chain for expanse/0.15.4/cpu
This dependency chain is intended to be used ONLY for supporting ORCA 4.2.1,
which is only distributed as a pre-compiled built against OpenMPI 3.1.4.
While the ORCA 4.2.1 binary may be compatible with OpenMPI 3.1.6, it'll
likely be safer to use the correct version over the long-term to avoid
any pitfalls we are not aware of at this time.