Add blast-plus@2.11.0 % gcc@10.2.0 to expanse/0.15.4/cpu
Note that the latest version of this Spack package [1] available from
the spack/spack develop branch was placed in the sdsc package repository
to override the default builtin version available in Spack v0.15.4
because blast-plus@2.9.0 failed to install with this older version of
the package. This package should be removed from the sdsc packge repo
when we upgrade to Spack v0.17.X in the future.

[1] commit id a8ccb8e116a1abfaa1d7ecb4e84f9f14ebab12ad