Add fastqc@0.11.7, fasttree@2.1.10, and fastx-toolkit@0.0.14
Note, however, these package spec install scripts have not yet been
included in the expanse/0.15.4/cpu dependency chain because there is no
record of any Expanse user requesting these software packages to be
installed. For example, the last request by a user for fastx-toolkit
was in 2017 when Gordon was being decomissioned --- the user wanted it
installed on Comet as part of the final transition.