Add gatk@4.1.8.1 % gcc@10.2.0 to expanse/0.15.4/cpu
A newer version of the gatk package from the spack/spack develop
branch [1] was placed in the sdsc package repo to override the default
builtin available in Spack v0.15.4. The primary reason for this was to
test if the latest versions of gatk no longer required openjdk@1.8.0 --
they still do. This package may be removed the the sdsc package repo
when we upgrade to Spack v0.17.X.

[1] commit id 418db4e9106371976d0ac3d95b4b4bf5adf3ffc8