Add raxml-ng@1.0.2 % gcc@10.2.0; installed after picard@2.20.8
Note that no Spack package existed for raxml-ng when the initial Spack
v0.15.4 intances were deployed in production. However, a new Spack
package was created by the Spack project in late 2020. Here, we've
placed the latest version of this Spack package from the spack/spack
develop branch into the sdsc package repository to replace the existing
custom one created by us. This package can be removed the the sdsc
package repo when we upgrade to Spack v0.17.X.