Add spades@3.15.2 % gcc@10.2.0; install after raxml-ng@1.0.2
Also placed latest version of spades package from spack/spack develop
branch in sdsc package repo and then modified it to include the latest
versions of spades, which were not yet in the package. These
modifications could be contributed back to the Spack project at some
point in the future.