Add stacks@2.3b % gcc@10.2.0 to expanse/0.15.4/cpu
Note, however, this package has not yet been included in the instance
package dependency chain because there is no record of an Expanse user
that has requested the package be installed.

In addition, it should be noted that the latest version of stacks@2.53
available in the current version of the Spack package failed to build
successfully on Expanse.