Add star@2.7.9a % gcc@10.2.0; installed after spades@3.15.2
Also placed latest version of star package from spack/spack develop
branch in sdsc package repo and then modified it to include the latest
versions of star, which was not yet in the package. This modification
could be contributed back to the Spack project at some point in the
future.