Add gmap-gsnap@2020-06-01 % gcc@10.2.0 to expanse/0.15.4/cpu
Note, however, this package has not yet been included in the instance
package dependency chain because there is no record of an Expanse user
that has requested the package be installed. Moreover, the package
failed to build successfully. We can revist later, if necessary. This
code is still under development. I would recommend we try and see if a
newer version will build successfully.