Add mafft@7.481 % gcc@10.2.0; install after jellyfish@2.2.7
Note, however, while this package was included in the instance package
dependency chain, there is no record of an Expanse user requesting the
package be installed. However, it is likely a package used by CIPRES and
they may have requested it to be installed via a private email to Mahidhar.

Also updated sdsc package repo with latest version of the package from
spack/spack develop branch. mafft@7.453 from the builtin package for
Spack v0.15.4 would not compile successfully. Note, however, this
version of mafft was compiled with gcc@9.2.0 in the original Expanse
deployment.