Add iq-tree@2.1.1 % gcc@10.2.0; install after salmon@1.4.0
Note, however, iq-tree is currently installed on Expanse using the AOCC
compiler. We should consider moving it back there when we upgrade to
Spack v0.17.X. In addition, there does not seem to be any strong reason
to rely on a custom Spack pacakge for iq-tree in the sdsc package
repository, other than adding additional versions. If these are the only
changes we require to the package, then we should contribute back these
trivial changes to the Spack project when we upgrade to v0.17.X.