Resolved uninstalled dependencies error with bcftools@1.10.2 install
The previous uninstalled dependencies error with libpng and zlib was
resolved by removing the explicit dependency on htslib@1.10.2. The
problem appears to have arisen during concetization because both htslib
and py-matplotlib have dependencies that rely on zlib, but the
concretizer attempts to only include a dependency once in the complete
spec prior to beginning the build. This uninstalled dependencies error
could be another manifestation of the cannot depend twice error, which
will hopefully be fixed in with the new concretizer in Spack v0.17.X.