Add gromacs@2019.6 % gcc@10.2.0; install after plumed@2.6.1
Note, this spec relies on the external LAPACK variant (+lapack), unlike
the current installations of GROMACS deployed on Expanse. Other GROMACS
dependencies already relied on OpenBLAS as their BLAS/LAPACK library.
The aim here with using OpenBLAS to satisfy the +lapack variant
dependency is to eliminate the use of multiple versions of LAPACK within
the same build of GROMACS.