Add gromacs@2020.4 % gcc@10.2.0; install after gromacs@2019.6
This spec did not include the patch/variant for plumed. When attempting
to concretize with the same dependencies as gromacs@2019.6 +plumed, it
failed. For some reason, gromacs@2020.4 wanted to utilize plumed@2.5.5
by default. We may want to look into this issue when we revisit writing
a custom GROMACS Spack package to correct a number of the shortcomings
of the existing official GROMACS Spack package. e.g., there should also
be an option on whether or not to use the internal FFTW; which CUDA
compute capabilities, etc.